📝 update payment documetations
