+++ 2NotyBackend/controllers/catalogs.js
@@ -178,80 +178,6 @@ const getMarks = async (req, res = response) => {
       });
   });
 };
-
-const getTypesPay = async (req, res = response) => {
-  pool.connect().then((client) => {
-    return client
-      .query(
-        `SELECT tp.id_tipo_pago,tp.tipo_pago,tp.id_estatus,e.estatus 
-      FROM tipo_pago tp, estatus e
-      WHERE tp.id_estatus=e.id_estatus;`
-      )
-      .then((response) => {
-        client.release();
-        res.status(200).json({
-          ok: true,
-          data: response.rows,
-        });
-      })
-      .catch((err) => {
-        client.release();
-        res.status(400).json({
-          ok: false,
-          msg: err,
-        });
-      });
-  });
-};
-const getPaymentsMeans = async (req, res = response) => {
-  pool.connect().then((client) => {
-    return client
-      .query(
-        `SELECT mp.id_medio_pago,mp.id_usuario,u.nombre,mp.id_tipo_pago,tp.tipo_pago,mp.numero_tarjeta_cuenta,mp.correo,to_char((mp.fecha_vigencia), 'DD/MM/YYYY')as fecha_vigencia 
-      FROM medios_pago mp, usuarios u, tipo_pago tp
-      WHERE mp.id_usuario=u.id_usuario AND mp.id_tipo_pago=tp.id_tipo_pago;`
-      )
-      .then((response) => {
-        client.release();
-        res.status(200).json({
-          ok: true,
-          data: response.rows,
-        });
-      })
-      .catch((err) => {
-        client.release();
-        res.status(400).json({
-          ok: false,
-          msg: err,
-        });
-      });
-  });
-};
-const getPaymentsUser = async (req, res = response) => {
-  pool.connect().then((client) => {
-    return client
-      .query(
-        `SELECT pu.id_pago_usuario,pu.id_usuario,u.nombre,pu.id_medio_pago,pu.monto_pago, to_char((pu.fecha_pago),'DD/MM/YYYY')as fecha_pago,pu.id_estatus,e.estatus
-      FROM pagos_usuarios pu,usuarios u, medios_pago mp, estatus e
-      WHERE pu.id_usuario=u.id_usuario AND pu.id_medio_pago=mp.id_medio_pago AND pu.id_estatus=e.id_estatus;`
-      )
-      .then((response) => {
-        client.release();
-        res.status(200).json({
-          ok: true,
-          data: response.rows,
-        });
-      })
-      .catch((err) => {
-        client.release();
-        res.status(400).json({
-          ok: false,
-          msg: err,
-        });
-      });
-  });
-};
-
 const updateState = async (req, res = response) => {
   const { id_pais, id_estatus, estado_provincia, id_estado } = req.body;
   const estadoUpperCase = estado_provincia.toUpperCase();
@@ -353,47 +279,7 @@ const updateProfiles = async (req, res = response) => {
       });
   });
 };
-const updateUser = async (req, res = response) => {
-  const {
-    id_usuario,
-    id_pais,
-    id_estado,
-    nombre,
-    correo,
-    id_estatus,
-    id_perfil,
-  } = req.body;
-  const nombreUpper = nombre.toUpperCase();
-  pool.connect().then((client) => {
-    return client
-      .query(
-        `UPDATE usuarios SET nombre=$2, correo=$3, id_estatus=$4, id_estado=$5, id_pais=$6, id_perfil=$7 where id_usuario=$1`,
-        [
-          id_usuario,
-          nombreUpper,
-          correo,
-          id_estatus,
-          id_estado,
-          id_pais,
-          id_perfil,
-        ]
-      )
-      .then((response) => {
-        client.release();
-        res.status(201).json({
-          ok: true,
-          data: response.command,
-        });
-      })
-      .catch((err) => {
-        client.release();
-        res.status(400).json({
-          ok: false,
-          msg: err,
-        });
-      });
-  });
-};
+
 const updateCompany = async (req, res = response) => {
   const { id_empresa, id_pais, empresa, razon_social, no_contrato } = req.body;
   const empresaUpper = empresa.toUpperCase();
@@ -455,34 +341,6 @@ const updateMark = async (req, res = response) => {
         id_empresa,
         marcaUpper,
       ])
-      .then((response) => {
-        client.release();
-        res.status(201).json({
-          ok: true,
-          data: response.command,
-        });
-      })
-      .catch((err) => {
-        client.release();
-        res.status(400).json({
-          ok: false,
-          msg: err,
-        });
-      });
-  });
-};
-
-const updateTypePay = async (req, res = response) => {
-  const { id_tipo_pago, tipo_pago, id_estatus } = req.body;
-  const tipopagoUpper = tipo_pago.toUpperCase();
-  pool.connect().then((client) => {
-    return client
-      .query(
-        `UPDATE tipo_pago
-      SET tipo_pago=$2, id_estatus=$3
-      WHERE id_tipo_pago=$1;`,
-        [id_tipo_pago, tipopagoUpper, id_estatus]
-      )
       .then((response) => {
         client.release();
         res.status(201).json({
@@ -857,110 +715,7 @@ const postMark = async (req, res = response) => {
   });
 };
 
-const postTypePay = async (req, res = response) => {
-  const { tipo_pago, id_estatus } = req.body;
-  const tipopagoUpper = tipo_pago.toUpperCase();
-  pool.connect().then((client) => {
-    return client
-      .query(`SELECT * FROM tipo_pago WHERE UPPER(tipo_pago)=$1`, [
-        tipopagoUpper,
-      ])
-      .then((response) => {
-        if (response.rows.length > 0) {
-          res.status(200).json({
-            ok: true,
-            msg: "Ya existe el tipo de pago",
-          });
-        } else {
-          return client
-            .query(
-              `INSERT INTO tipo_pago (
-              tipo_pago, id_estatus)
-              VALUES ($1, $2);`,
-              [tipopagoUpper, id_estatus]
-            )
-            .then((response) => {
-              client.release();
-              res.status(201).json({
-                ok: true,
-                msg: response.command,
-              });
-            })
-            .catch((err) => {
-              client.release();
-              res.status(400).json({
-                ok: false,
-                msg: err,
-              });
-            });
-        }
-      })
-      .catch((err) => {
-        client.release();
-        res.status(400).json({
-          ok: false,
-          msg: err,
-        });
-      });
-  });
-};
-const postPaymentsMeans = async (req, res = response) => {
-  const {
-    id_usuario,
-    id_tipo_pago,
-    numero_tarjeta_cuenta,
-    correo,
-    fecha_vigencia,
-  } = req.body;
-  pool.connect().then((client) => {
-    // return client
-    //   .query(`SELECT * FROM medios_pago WHERE UPPER(tipo_pago)=$1`,
-    //     [tipopagoUpper])
-    //   .then((response) => {
-    //     if (response.rows.length > 0) {
-    //       res.status(200).json({
-    //         ok: true,
-    //         msg: "Ya existe el tipo de pago",
-    //       });
-    //     } else {
-    return client
-      .query(
-        `INSERT INTO medios_pago (
-              id_usuario,id_tipo_pago,numero_tarjeta_cuenta,correo,fecha_vigencia)
-              VALUES ($1, $2, $3, $4, $5);`,
-        [
-          id_usuario,
-          id_tipo_pago,
-          numero_tarjeta_cuenta,
-          correo,
-          fecha_vigencia,
-        ]
-      )
-      .then((response) => {
-        client.release();
-        res.status(201).json({
-          ok: true,
-          msg: response.command,
-        });
-      })
-      .catch((err) => {
-        client.release();
-        res.status(400).json({
-          ok: false,
-          msg: err,
-        });
-      });
-    //   }
-    // })
-    // .catch((err) => {
-    //   client.release();
-    //   res.status(400).json({
-    //     ok: false,
-    //     msg: err,
-    //   });
-    // });
-  });
-};
+
 
 module.exports = {
   getStatus,
@@ -971,9 +726,6 @@ module.exports = {
   getBills,
   getCategories,
   getMarks,
-  getTypesPay,
-  getPaymentsUser,
-  getPaymentsMeans,
   postStates,
   postContry,
   postStatus,
@@ -981,17 +733,13 @@ module.exports = {
   postCompany,
   postCategory,
   postMark,
-  postTypePay,
-  postPaymentsMeans,
   updateState,
   updateCountry,
   updateStatus,
   updateProfiles,
-  updateUser,
   updateCompany,
   updateCategory,
   updateMark,
-  updateTypePay,
   activateCountry,
   activateState,
 };
