+++ 2NotyBackend/routes/catalogs.js
@@ -23,7 +23,6 @@ const {
   getCompanies,
   updateCompany,
   postCompany,
-  updateUser,
   getBills,
   getCategories,
   updateCategory,
@@ -31,12 +30,6 @@ const {
   getMarks,
   updateMark,
   postMark,
-  getTypesPay,
-  postTypePay,
-  updateTypePay,
-  getPaymentsMeans,
-  postPaymentsMeans,
-  getPaymentsUser,
 } = require("../controllers/catalogs");
 router.get("/getStatus", getStatus);
 router.get("/getCountries", getCountries);
@@ -46,9 +39,6 @@ router.get("/getCompanies", getCompanies);
 router.get("/getBills", getBills);
 router.get("/getCategories", getCategories);
 router.get("/getMarks", getMarks);
-router.get("/getTypespay", getTypesPay);
-router.get("/getPaymentsmeans", getPaymentsMeans);
-router.get("/getPaymentsuser", getPaymentsUser);
 
 router.post(
   "/state",
@@ -88,20 +78,16 @@ router.post(
 );
 router.post("/category", postCategory);
 router.post("/mark", postMark);
-
-router.post("/typepay", postTypePay);
-router.post("/paymentsmeans", postPaymentsMeans);
-
 router.put("/updateCountry", updateCountry);
 router.put("/updateState", updateState);
 router.put("/updateStatus", updateStatus);
 router.put("/updateProfile", updateProfiles);
-router.put("/updateUser", updateUser);
+
 router.put("/updateCompany", updateCompany);
 router.put("/updateCategory", updateCategory);
 router.put("/updateMark", updateMark);
 
-router.put("/updateTypepay", updateTypePay);
+
 /* router.put("/activateCountry/:id", activateCountry);
 router.put("/activateState/:id", activateState); */
